Refactor a bit
