+++ nucliadb/src/nucliadb/writer/api/v1/export_import.py
@@ -45,7 +45,7 @@ from nucliadb.writer import logger
 from nucliadb.writer.api.utils import only_for_onprem
 from nucliadb.writer.api.v1.knowledgebox import create_kb
 from nucliadb.writer.api.v1.router import KB_PREFIX, KBS_PREFIX, api
-from nucliadb.writer.back_pressure import maybe_back_pressure
+from nucliadb.common.back_pressure import maybe_back_pressure
 from nucliadb_models.export_import import (
     CreateExportResponse,
     CreateImportResponse,
