+++ nucliadb/src/nucliadb/writer/api/v1/field.py
@@ -40,7 +40,7 @@ from nucliadb.writer.api.v1.resource import (
     validate_rid_exists_or_raise_error,
 )
 from nucliadb.writer.api.v1.router import KB_PREFIX, RESOURCE_PREFIX, RSLUG_PREFIX, api
-from nucliadb.writer.back_pressure import maybe_back_pressure
+from nucliadb.common.back_pressure import maybe_back_pressure
 from nucliadb.writer.resource.audit import parse_audit
 from nucliadb.writer.resource.field import (
     ResourceClassifications,
